requirements.txt
requests>=2.28.0
pandas>=1.5.0
pyarrow>=10.0.0
python-dotenv>=0.21.0

# Instalar
pip install -r requirements.txt